add new typeChecker listener, refactor structureBuilder to collect defined class names only
